server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.2
numpy==1.26.4
pydantic==2.9.2
python-dotenv==1.0.1
anthropic==0.39.0
openai==1.47.0
httpx==0.27.2
python-multipart==0.0.9
